Error fixing and new method
Fixed the bug "Primary Key Attribute is not a supported dictionary key using converter of type 'System.Text.Json.Serialization.Converters.SmallObjectWithParameterizedConstructorConverter'" by installing the Microsoft.AspNetCore.Mvc.NewtonsoftJson library, similarly to the projects service, part of the NexusPilot eco system.
Introduced a method for deleting tasks in the TaskService. In addition to that, introduced DeletionController, that is the enpoint for deletion opperations.