Introduced MutationController and some additional adjustments
Introduced a method for changing the "done" role in database to true, when called. The MarkTaskAsDone method is introduced, so when called, a task is marked as done. The method is called through the MutationController. The MutationController's main purpose is to handle all mutations on the task entity.
Additionally, the TaskItem model was modified. The model now includes StartDate, EndDate and Done members, The change was necessary, in order the model to match the updated database entity.
As the TaskItem model was changed, the CreationController was also modified. The schema for creating a TaskItem, now includes StartDate and EndDate members.
Introduced a custom exception, to be thrown, when the requested recourse was not present in the database.